requirements.txt
Django>=4.2
djangorestframework>=3.14
python-decouple>=3.8
PyMuPDF>=1.23.22
cohere>=4.35
djangorestframework-simplejwt>=5.3.1
